include: linux: pwm: Convert phase to u64
This change is a follow-up to
commit 4fa760fa62a9 ("pwm: Convert period and duty cycle to u64"),
converting the remaining parameters to u64.

(cherry picked from commit 53239d5626c906917634f961facd4fbf14ecf060)

+++ include/linux/pwm.h
@@ -41,7 +41,7 @@ enum pwm_polarity {
  */
 struct pwm_args {
 	u64 period;
-	unsigned int phase;
+	u64 phase;
 	enum pwm_polarity polarity;
 };
 
@@ -65,7 +65,7 @@ enum {
 struct pwm_state {
 	u64 period;
 	u64 duty_cycle;
-	unsigned int phase;
+	u64 phase;
 	enum pwm_polarity polarity;
 	bool enabled;
 	bool usage_power;
@@ -151,13 +151,13 @@ static inline u64 pwm_get_duty_cycle(const struct pwm_device *pwm)
 	return state.duty_cycle;
 }
 
-static inline void pwm_set_phase(struct pwm_device *pwm, unsigned int phase)
+static inline void pwm_set_phase(struct pwm_device *pwm, u64 phase)
 {
 	if (pwm)
 		pwm->state.phase = phase;
 }
 
-static inline unsigned int pwm_get_phase(const struct pwm_device *pwm)
+static inline u64 pwm_get_phase(const struct pwm_device *pwm)
 {
 	struct pwm_state state;
 
@@ -340,7 +340,7 @@ struct pwm_chip {
 struct pwm_capture {
 	u64 period;
 	u64 duty_cycle;
-	unsigned int phase;
+	u64 phase;
 };
 
 #if IS_ENABLED(CONFIG_PWM)
